[FEAT] Calculate ReadyAt Time for Lava Pit
